# packages/shared-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Video Digestor Phase 3 API
  version: 0.1.0
  description: API contract for subscriptions, ingest polling, jobs, videos and artifacts.
servers:
  - url: http://127.0.0.1:8000
paths:
  /api/v1/subscriptions:
    get:
      operationId: listSubscriptions
      parameters:
        - in: query
          name: platform
          schema:
            $ref: '#/components/schemas/Platform'
        - in: query
          name: enabled_only
          schema:
            type: boolean
      responses:
        '200':
          description: Subscription list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
    post:
      operationId: upsertSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriptionUpsertRequest'
      responses:
        '200':
          description: Upsert result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionUpsertResponse'
  /api/v1/subscriptions/{id}:
    delete:
      operationId: removeSubscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
...
components:
  schemas:
    Platform:
      type: string
      enum: [bilibili, youtube]
    SourceType:
      type: string
      enum: [bilibili_uid, youtube_channel_id, url]
...
    Subscription:
      type: object
      required: [id, platform, source_type, source_value, rsshub_route, enabled, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        platform:
          $ref: '#/components/schemas/Platform'
        source_type:
          $ref: '#/components/schemas/SourceType'
        source_value:
          type: string
        rsshub_route:
          type: string
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SubscriptionUpsertRequest:
      type: object
      required: [platform, source_type, source_value]
      properties:
        platform:
          $ref: '#/components/schemas/Platform'
        source_type:
          $ref: '#/components/schemas/SourceType'
        source_value:
          type: string
        rsshub_route:
          type: string
        enabled:
          type: boolean
    SubscriptionUpsertResponse:
      type: object
      required: [subscription, created]
      properties:
        subscription:
          $ref: '#/components/schemas/Subscription'
        created:
          type: boolean
...